+++ webpack.config.js
@@ -3,6 +3,7 @@
 'use strict';
 
 const path = require('path');
+const CopyPlugin = require('copy-webpack-plugin');
 
 /**@type {import('webpack').Configuration}*/
 const config = {
@@ -41,6 +42,13 @@ const config = {
         ]
       }
     ]
-  }
+  },
+  plugins: [
+    new CopyPlugin({
+      patterns: [
+        { from: 'voice-list.json', to: 'voice-list.json' }
+      ],
+    }),
+  ]
 };
 module.exports = config;